requirements.txt
asgiref==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
get==2019.4.13
graphviz==0.14
idna==2.8
importlib-metadata==1.6.1
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
Markdown==3.2
MarkupSafe==1.1.1
matplotlib==3.2.2
more-itertools==8.4.0
nose==1.3.7
numpy==1.18.1
packaging==20.4
pandas==0.25.3
pluggy==0.13.1
post==2019.4.13
public==2019.4.13
py==1.8.2
pycparser==2.19
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.2.4
zipp==3.1.0
